test 파일을 분할하고, custom type 예제 작성을 준비 중. wrapper코드 전체에 namespace를 둬야겠다.

+++ msgpack_wrapper/MapReader.h
@@ -25,8 +25,8 @@ public:
 
 private:
 	Reader& Ready(const std::string& key);
-	Reader& ReadyKey(int index);
-	Reader& ReadyValue(int index);
+	Reader& ReadyKey(size_t index);
+	Reader& ReadyValue(size_t index);
 
 	class Impl;
 	Impl* impl_;
